requirements.txt
langchain==0.3.27
langchain-groq==0.3.6
langchain-community==0.3.27
langchain-core==0.3.72
langgraph==0.6.4
faiss-cpu==1.11.0.post1
streamlit==1.49.1
tavily-python==0.7.10
langsmith==0.4.8
PyYAML==6.0.2
python-box==7.3.2
